Fixed sensor tests
